+++ packages/frameworks/react-mdx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qualcomm-ui/react-mdx Changelog
 
+## 2.9.8
+
+Jun 29th, 2026
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/react@1.25.0]
+
 ## 2.9.7
 
 Jun 10th, 2026